src/Realtime_Object_Detector/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.17
Werkzeug==1.0.1
mss==2.0.22
numpy==1.18.4
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.1.2
Keras==2.2.5
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==6.2.0
tensorflow-gpu==1.15.0
progressbar2
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
tqdm==4.28.1
